+++ serviceconnector/src/deleteLinker/IPickLinkerContext.ts
@@ -3,7 +3,7 @@
 *  Licensed under the MIT License. See License.md in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 
-import { ICreateLinkerContext } from "../createServiceConnector/ICreateLinkerContext";
+import { ICreateLinkerContext } from "../createLinker/ICreateLinkerContext";
 
 export interface IPickLinkerContext extends ICreateLinkerContext {
     linkerName?: string;
