+++ serviceconnector/src/index.ts
@@ -3,12 +3,12 @@
 *  Licensed under the MIT License. See License.md in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 export { KnownClientType, LinkerResource } from "@azure/arm-servicelinker";
-export * from './createServiceConnector/ICreateLinkerContext';
-export * from './createServiceConnector/LinkerCreateStep';
-export * from './createServiceConnector/LinkerNameStep';
-export * from './createServiceConnector/TargetServiceListStep';
-export * from './createServiceConnector/createLinker';
-export * from './deleteServiceConnector/IPickLinkerContext';
-export * from './deleteServiceConnector/deleteLinker';
-export * from './validateServiceConnector/validateLinker';
+export * from './createLinker/ICreateLinkerContext';
+export * from './createLinker/LinkerCreateStep';
+export * from './createLinker/LinkerNameStep';
+export * from './createLinker/TargetServiceListStep';
+export * from './createLinker/createLinker';
+export * from './deleteLinker/IPickLinkerContext';
+export * from './deleteLinker/deleteLinker';
+export * from './validateLinker/validateLinker';
